[Mailer][FrameworkBundle] Add MailerBundle
The mailer services move to the Mailer component, which now provides its own bundle and
the "mailer" configuration. The "framework.mailer" configuration becomes an alias of it.

A new RemoveMissingDependenciesPass takes over the decisions the extension used to make by
reading other sections: the data collector is registered in debug mode and dropped when the
profiler is disabled, and the message logger listener is gated on the profiler state, or
dropped when neither the profiler nor the test client collects the messages it keeps.

On the FrameworkBundle side, the notifier email channel cannot read the mailer envelope
sender at load time anymore, so RemoveMissingMailerDependenciesPass forwards it, and drops
the channel when the mailer is missing.
