+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -24,7 +24,6 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\Cookie;
 use Symfony\Component\HttpFoundation\IpUtils;
-use Symfony\Component\Mailer\Mailer;
 use Symfony\Component\Notifier\Notifier;
 use Symfony\Component\Serializer\Encoder\JsonDecode;
 use Symfony\Component\Serializer\Serializer;
@@ -165,7 +164,7 @@ class Configuration implements ConfigurationInterface
         $this->addSchedulerSection($rootNode);
         $this->addRobotsIndexSection($rootNode);
         $this->addHttpClientSection($rootNode);
-        $this->addMailerSection($rootNode, $enableIfStandalone);
+        $this->addMailerSection($rootNode);
         $this->addSecretsSection($rootNode);
         $this->addNotifierSection($rootNode, $enableIfStandalone);
         $this->addRateLimiterSection($rootNode);
@@ -1135,271 +1134,14 @@ class Configuration implements ConfigurationInterface
         ;
     }
 
-    /**
-     * @param-immediately-invoked-callable $enableIfStandalone
-     */
-    private function addMailerSection(ArrayNodeDefinition $rootNode, callable $enableIfStandalone): void
+    private function addMailerSection(ArrayNodeDefinition $rootNode): void
     {
         $rootNode
             ->children()
-                ->arrayNode('mailer')
-                    ->info('Mailer configuration')
-                    ->{$enableIfStandalone('symfony/mailer', Mailer::class)}()
-                    ->validate()
-                        ->ifTrue(static fn ($v) => isset($v['dsn']) && $v['transports'])
-                        ->thenInvalid('"dsn" and "transports" cannot be used together.')
-                    ->end()
-                    ->children()
-                        ->scalarNode('message_bus')->defaultNull()->info('The message bus to use. Defaults to the default bus if the Messenger component is installed.')->end()
-                        ->scalarNode('dsn')->defaultNull()->end()
-                        ->arrayNode('transports', 'transport')
-                            ->useAttributeAsKey('name')
-                            ->arrayPrototype()
-                                ->beforeNormalization()
-                                    ->ifString()
-                                    ->then(static fn (string $dsn) => ['dsn' => $dsn])
-                                ->end()
-                                ->children()
-                                    ->scalarNode('dsn')->end()
-                                    ->scalarNode('rate_limiter')
-                                        ->defaultNull()
-                                        ->info('Rate limiter name used to limit the number of messages sent through this transport; when the limit is exceeded, sending fails with a RateLimitExceededException.')
-                                    ->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('envelope')
-                            ->info('Mailer Envelope configuration')
-                            ->children()
-                                ->scalarNode('sender')->end()
-                                ->arrayNode('recipients', 'recipient')
-                                    ->performNoDeepMerging()
-                                    ->acceptAndWrap(['string'])
-                                    ->beforeNormalization()
-                                        ->ifArray()
-                                        ->then(static fn ($v) => array_values(array_filter($v)))
-                                    ->end()
-                                    ->prototype('scalar')->end()
-                                ->end()
-                                ->arrayNode('allowed_recipients', 'allowed_recipient')
-                                    ->info('A list of regular expressions that allow recipients when "recipients" option is defined.')
-                                    ->example(['.*@example\.com'])
-                                    ->performNoDeepMerging()
-                                    ->acceptAndWrap(['string'])
-                                    ->beforeNormalization()
-                                        ->ifArray()
-                                        ->then(static fn ($v) => array_values(array_filter($v)))
-                                    ->end()
-                                    ->prototype('scalar')->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('headers', 'header')
-                            ->normalizeKeys(false)
-                            ->useAttributeAsKey('name')
-                            ->prototype('array')
-                                ->normalizeKeys(false)
-                                ->acceptAndWrap(['string'], 'value')
-                                ->beforeNormalization()
-                                    ->ifArray()
-                                    ->then(static fn ($v) => array_keys($v) !== ['value'] ? ['value' => $v] : $v)
-                                ->end()
-                                ->children()
-                                    ->variableNode('value')->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('tracking')
-                            ->info('Default open/click tracking for every message that does not carry an explicit "X-Track" header; null keeps each provider\'s default. An "X-Track" entry in the "headers" option wins over this one.')
-                            ->addDefaultsIfNotSet()
-                            ->children()
-                                ->enumNode('opens')->values([true, false, null])->defaultNull()->end()
-                                ->enumNode('clicks')->values([true, false, null])->defaultNull()->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('dkim_signer')
-                            ->addDefaultsIfNotSet()
-                            ->canBeEnabled()
-                            ->info('DKIM signer configuration')
-                            ->children()
-                                ->scalarNode('key')
-                                    ->info('Key content, or path to key (in PEM format with the `file://` prefix)')
-                                    ->defaultValue('')
-                                    ->cannotBeEmpty()
-                                ->end()
-                                ->scalarNode('domain')->defaultValue('')->end()
-                                ->scalarNode('select')->defaultValue('')->end()
-                                ->scalarNode('passphrase')
-                                    ->info('The private key passphrase')
-                                    ->defaultValue('')
-                                ->end()
-                                ->arrayNode('options', 'option')
-                                    ->performNoDeepMerging()
-                                    ->normalizeKeys(false)
-                                    ->useAttributeAsKey('name')
-                                    ->prototype('variable')->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('smime_signer')
-                            ->addDefaultsIfNotSet()
-                            ->canBeEnabled()
-                            ->info('S/MIME signer configuration')
-                            ->children()
-                                ->scalarNode('key')
-                                    ->info('Path to key (in PEM format)')
-                                    ->defaultValue('')
-                                    ->cannotBeEmpty()
-                                ->end()
-                                ->scalarNode('certificate')
-                                    ->info('Path to certificate (in PEM format without the `file://` prefix)')
-                                    ->defaultValue('')
-                                    ->cannotBeEmpty()
-                                ->end()
-                                ->scalarNode('passphrase')
-                                    ->info('The private key passphrase')
-                                    ->defaultNull()
-                                ->end()
-                                ->scalarNode('extra_certificates')->defaultNull()->end()
-                                ->integerNode('sign_options')->defaultNull()->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('smime_encrypter')
-                            ->addDefaultsIfNotSet()
-                            ->canBeEnabled()
-                            ->info('S/MIME encrypter configuration')
-                            ->validate()
-                                ->ifTrue(static fn ($v) => $v['enabled'] && '' === $v['repository'] && !$v['certificates'])
-                                ->thenInvalid('You must configure either "smime_encrypter.repository" or "smime_encrypter.certificates".')
-                            ->end()
-                            ->validate()
-                                ->ifTrue(static fn ($v) => '' !== $v['repository'] && $v['certificates'])
-                                ->thenInvalid('You cannot use both "smime_encrypter.repository" and "smime_encrypter.certificates" at the same time.')
-                            ->end()
-                            ->children()
-                                ->scalarNode('repository')
-                                    ->info('S/MIME certificate repository service. This service shall implement the `Symfony\Component\Mailer\EventListener\SmimeCertificateRepositoryInterface`.')
-                                    ->defaultValue('')
-                                ->end()
-                                ->arrayNode('certificates')
-                                    ->info('Recipient S/MIME certificates, as a map of email address to certificate file path. Cannot be used together with "repository".')
-                                    ->useAttributeAsKey('email')
-                                    ->scalarPrototype()->end()
-                                ->end()
-                                ->enumNode('on_missing_certificate')
-                                    ->info('Default behavior when a recipient has no S/MIME certificate: "send_unencrypted" (send the message unencrypted, deprecated since 8.2), "fail" (throw an exception), "encrypt" (encrypt for the recipients that have a certificate, the others receive an unreadable message), "skip" (encrypt for the recipients that have a certificate and drop the others from the envelope; note that "framework.mailer.envelope.recipients" is applied afterwards and overrides that list). Can be overridden per message by setting the "X-SMime-Encrypt" header to one of these values.')
-                                    ->values(['send_unencrypted', 'fail', 'encrypt', 'skip'])
-                                    ->defaultValue('send_unencrypted')
-                                ->end()
-                                ->booleanNode('encrypt_for_sender')
-                                    ->info('Also encrypt for the sender, when a certificate is available for its address, so that the sender can read the messages it sent.')
-                                    ->defaultFalse()
-                                ->end()
-                                ->integerNode('cipher')
-                                    ->info('A set of algorithms used to encrypt the message')
-                                    ->defaultNull()
-                                    ->beforeNormalization()
-                                        ->ifString()
-                                        ->then(static function ($v): int {
-                                            $ciphers = self::getOpensslCiphers();
-
-                                            if (!isset($ciphers[$v])) {
-                                                throw new \InvalidArgumentException(\sprintf('"%s" is not a valid OPENSSL cipher.', $v));
-                                            }
-
-                                            return $ciphers[$v];
-                                        })
-                                    ->end()
-                                    ->validate()
-                                        ->ifTrue(static fn ($v) => null !== $v && ($ciphers = self::getOpensslCiphers()) && !\in_array($v, $ciphers, true))
-                                        ->thenInvalid('You must provide a valid cipher.')
-                                    ->end()
-                                ->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('pgp_signer')
-                            ->addDefaultsIfNotSet()
-                            ->canBeEnabled()
-                            ->info('PGP/MIME signer configuration')
-                            ->validate()
-                                ->ifTrue(static fn ($v) => $v['enabled'] && '' === $v['secret_key'])
-                                ->thenInvalid('You must configure "pgp_signer.secret_key".')
-                            ->end()
-                            ->children()
-                                ->scalarNode('secret_key')
-                                    ->info('Path to the secret key (ASCII armored format without the `file://` prefix)')
-                                    ->defaultValue('')
-                                ->end()
-                                ->scalarNode('public_key')
-                                    ->info('Path to the public key (ASCII armored format without the `file://` prefix)')
-                                    ->defaultNull()
-                                ->end()
-                                ->scalarNode('passphrase')
-                                    ->info('The secret key passphrase')
-                                    ->defaultNull()
-                                ->end()
-                                ->scalarNode('binary')
-                                    ->info('Path to the GnuPG binary')
-                                    ->defaultValue('gpg')
-                                ->end()
-                                ->enumNode('digest_algorithm')
-                                    ->info('The digest algorithm used to sign the message')
-                                    ->values(['SHA224', 'SHA256', 'SHA384', 'SHA512'])
-                                    ->defaultValue('SHA512')
-                                ->end()
-                            ->end()
-                        ->end()
-                        ->arrayNode('pgp_encrypter')
-                            ->addDefaultsIfNotSet()
-                            ->canBeEnabled()
-                            ->info('PGP/MIME encrypter configuration')
-                            ->validate()
-                                ->ifTrue(static fn ($v) => $v['enabled'] && '' === $v['repository'] && !$v['keys'])
-                                ->thenInvalid('You must configure either "pgp_encrypter.repository" or "pgp_encrypter.keys".')
-                            ->end()
-                            ->validate()
-                                ->ifTrue(static fn ($v) => '' !== $v['repository'] && $v['keys'])
-                                ->thenInvalid('You cannot use both "pgp_encrypter.repository" and "pgp_encrypter.keys" at the same time.')
-                            ->end()
-                            ->children()
-                                ->scalarNode('repository')
-                                    ->info('Service or class implementing `Symfony\Component\Mailer\EventListener\PgpPublicKeyRepositoryInterface` to provide recipient PGP public keys.')
-                                    ->defaultValue('')
-                                ->end()
-                                ->arrayNode('keys')
-                                    ->info('Recipient PGP public keys, as a map of email address to public key file path. Cannot be used together with "repository".')
-                                    ->useAttributeAsKey('email')
-                                    ->scalarPrototype()->end()
-                                ->end()
-                                ->scalarNode('binary')
-                                    ->info('Path to the GnuPG binary')
-                                    ->defaultValue('gpg')
-                                ->end()
-                                ->enumNode('cipher_algorithm')
-                                    ->info('The cipher algorithm used to encrypt the message')
-                                    ->values(['AES', 'AES192', 'AES256', 'TWOFISH', 'CAMELLIA128', 'CAMELLIA192', 'CAMELLIA256'])
-                                    ->defaultValue('AES256')
-                                ->end()
-                                ->floatNode('timeout')
-                                    ->info('Timeout in seconds for the GPG process (null for no timeout)')
-                                    ->defaultValue(60.0)
-                                ->end()
-                                ->booleanNode('hide_recipients')
-                                    ->info('Hide every recipient\'s key ID in the encrypted message (gpg --hidden-recipient). Recipients listed in the Bcc header are always hidden regardless of this option; set it to true to also hide To and Cc recipients.')
-                                    ->defaultFalse()
-                                ->end()
-                                ->enumNode('on_missing_key')
-                                    ->info('Default behavior when a recipient has no PGP public key: "fail" (throw an exception), "encrypt" (encrypt for the recipients that have a key, the others receive an unreadable message), "skip" (encrypt for the recipients that have a key and drop the others from the envelope; note that "framework.mailer.envelope.recipients" is applied afterwards and overrides that list). The message is never sent unencrypted. Can be overridden per message by setting the "X-Pgp-Encrypt" header to one of these values.')
-                                    ->values(['fail', 'encrypt', 'skip'])
-                                    ->defaultValue('fail')
-                                ->end()
-                                ->booleanNode('encrypt_for_sender')
-                                    ->info('Also encrypt for the sender, when a public key is available for its address, so that the sender can read the messages it sent.')
-                                    ->defaultFalse()
-                                ->end()
-                            ->end()
-                        ->end()
-                    ->end()
+                ->variableNode('mailer')
+                    ->aliasOf('mailer')
+                    ->treatFalseLike(['enabled' => false])
+                    ->treatTrueLike(['enabled' => true])
                 ->end()
             ->end()
         ;
@@ -1523,21 +1265,5 @@ class Configuration implements ConfigurationInterface
                 ->end()
             ->end()
         ;
-    }
-
-    /**
-     * @return array<string, int> The values of the OPENSSL_CIPHER_* constants, keyed by their unprefixed name
-     */
-    private static function getOpensslCiphers(): array
-    {
-        $ciphers = [];
-
-        foreach (get_defined_constants(true)['openssl'] ?? [] as $name => $value) {
-            if (str_starts_with($name, 'OPENSSL_CIPHER_')) {
-                $ciphers[substr($name, \strlen('OPENSSL_CIPHER_'))] = $value;
-            }
-        }
-
-        return $ciphers;
     }
 }
