+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -14,14 +14,12 @@ namespace Symfony\Bundle\FrameworkBundle\Tests\DependencyInjection;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\IgnoreDeprecations;
-use PHPUnit\Framework\Attributes\RequiresPhpExtension;
 use PHPUnit\Framework\Attributes\TestWith;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bundle\FrameworkBundle\DependencyInjection\Configuration;
 use Symfony\Bundle\FullStack;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
 use Symfony\Component\Config\Definition\Processor;
-use Symfony\Component\Mailer\Mailer;
 use Symfony\Component\Notifier\Notifier;
 use Symfony\Component\Serializer\Encoder\JsonDecode;
 
@@ -450,78 +448,6 @@ class ConfigurationTest extends TestCase
         }
     }
 
-    #[RequiresPhpExtension('openssl')]
-    public function testMailerSmimeEncrypterCipherAsConstantName()
-    {
-        $config = self::processSmimeEncrypterConfig(['cipher' => 'AES_256_CBC']);
-        $this->assertSame(\OPENSSL_CIPHER_AES_256_CBC, $config['mailer']['smime_encrypter']['cipher']);
-
-        $config = self::processSmimeEncrypterConfig(['cipher' => 'RC2_40']);
-        $this->assertSame(\OPENSSL_CIPHER_RC2_40, $config['mailer']['smime_encrypter']['cipher']);
-    }
-
-    #[RequiresPhpExtension('openssl')]
-    public function testMailerSmimeEncrypterCipherAsConstantValue()
-    {
-        $config = self::processSmimeEncrypterConfig(['cipher' => \OPENSSL_CIPHER_AES_256_CBC]);
-        $this->assertSame(\OPENSSL_CIPHER_AES_256_CBC, $config['mailer']['smime_encrypter']['cipher']);
-
-        $config = self::processSmimeEncrypterConfig(['cipher' => \OPENSSL_CIPHER_RC2_40]);
-        $this->assertSame(\OPENSSL_CIPHER_RC2_40, $config['mailer']['smime_encrypter']['cipher']);
-    }
-
-    public function testMailerSmimeEncrypterCipherDefaultsToNull()
-    {
-        $config = self::processSmimeEncrypterConfig([]);
-
-        $this->assertNull($config['mailer']['smime_encrypter']['cipher']);
-    }
-
-    public function testMailerSmimeEncrypterCipherAsInvalidConstantName()
-    {
-        $this->expectException(\InvalidArgumentException::class);
-        $this->expectExceptionMessage('"NOT_A_CIPHER" is not a valid OPENSSL cipher.');
-
-        self::processSmimeEncrypterConfig(['cipher' => 'NOT_A_CIPHER']);
-    }
-
-    #[RequiresPhpExtension('openssl')]
-    public function testMailerSmimeEncrypterCipherAsInvalidConstantValue()
-    {
-        $this->expectException(InvalidConfigurationException::class);
-        $this->expectExceptionMessage('Invalid configuration for path "framework.mailer.smime_encrypter.cipher": You must provide a valid cipher.');
-
-        self::processSmimeEncrypterConfig(['cipher' => 123456]);
-    }
-
-    public function testMailerSmimeEncrypterCipherIsNotValidatedWithoutOpenssl()
-    {
-        if (\extension_loaded('openssl')) {
-            $this->markTestSkipped('The "openssl" extension is loaded.');
-        }
-
-        $config = self::processSmimeEncrypterConfig(['cipher' => 123456]);
-        $this->assertSame(123456, $config['mailer']['smime_encrypter']['cipher']);
-
-        $config = self::processSmimeEncrypterConfig([]);
-        $this->assertNull($config['mailer']['smime_encrypter']['cipher']);
-    }
-
-    private static function processSmimeEncrypterConfig(array $smimeEncrypter): array
-    {
-        return (new Processor())->processConfiguration(new Configuration(true), [
-            [
-                'http_method_override' => false,
-                'handle_all_throwables' => true,
-                'php_errors' => ['log' => true],
-                'mailer' => [
-                    'dsn' => 'null://null',
-                    'smime_encrypter' => $smimeEncrypter + ['repository' => 'my_certificate_repository'],
-                ],
-            ],
-        ]);
-    }
-
     protected static function getBundleDefaultConfig()
     {
         return [
@@ -653,60 +579,6 @@ class ConfigurationTest extends TestCase
                 'throw' => true,
             ],
             'disallow_search_engine_index' => true,
-            'mailer' => [
-                'dsn' => null,
-                'transports' => [],
-                'enabled' => !class_exists(FullStack::class) && class_exists(Mailer::class),
-                'message_bus' => null,
-                'headers' => [],
-                'tracking' => [
-                    'opens' => null,
-                    'clicks' => null,
-                ],
-                'dkim_signer' => [
-                    'enabled' => false,
-                    'options' => [],
-                    'key' => '',
-                    'domain' => '',
-                    'select' => '',
-                    'passphrase' => '',
-                ],
-                'smime_signer' => [
-                    'enabled' => false,
-                    'key' => '',
-                    'certificate' => '',
-                    'passphrase' => null,
-                    'extra_certificates' => null,
-                    'sign_options' => null,
-                ],
-                'smime_encrypter' => [
-                    'enabled' => false,
-                    'repository' => '',
-                    'certificates' => [],
-                    'on_missing_certificate' => 'send_unencrypted',
-                    'encrypt_for_sender' => false,
-                    'cipher' => null,
-                ],
-                'pgp_signer' => [
-                    'enabled' => false,
-                    'secret_key' => '',
-                    'public_key' => null,
-                    'passphrase' => null,
-                    'binary' => 'gpg',
-                    'digest_algorithm' => 'SHA512',
-                ],
-                'pgp_encrypter' => [
-                    'enabled' => false,
-                    'repository' => '',
-                    'keys' => [],
-                    'binary' => 'gpg',
-                    'cipher_algorithm' => 'AES256',
-                    'timeout' => 60.0,
-                    'hide_recipients' => false,
-                    'on_missing_key' => 'fail',
-                    'encrypt_for_sender' => false,
-                ],
-            ],
             'notifier' => [
                 'enabled' => !class_exists(FullStack::class) && class_exists(Notifier::class),
                 'message_bus' => null,
