--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer.php
@@ -1,16 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-        'envelope' => [
-            'sender' => 'sender@example.org',
-            'recipients' => ['redirected@example.org', 'redirected1@example.org'],
-        ],
-        'headers' => [
-            'from' => 'from@example.org',
-            'bcc' => ['bcc1@example.org', 'bcc2@example.org'],
-            'foo' => 'bar',
-        ],
-    ],
-]);
