--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer_with_pgp_without_source.php
@@ -1,10 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-        'pgp_encrypter' => [
-            'enabled' => true,
-        ],
-    ],
-]);
