--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer_with_smime_repository.php
@@ -1,16 +1,0 @@
-<?php
-
-use Symfony\Component\Mailer\EventListener\InMemorySmimeCertificateRepository;
-
-$container->register('my_repository', InMemorySmimeCertificateRepository::class)
-    ->setArguments([[]]);
-
-$container->loadFromExtension('framework', [
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-        'smime_encrypter' => [
-            'enabled' => true,
-            'repository' => 'my_repository',
-        ],
-    ],
-]);
