--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer_with_smime_repository_and_certificates.php
@@ -1,14 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-        'smime_encrypter' => [
-            'enabled' => true,
-            'repository' => 'my_repository',
-            'certificates' => [
-                'r1@example.com' => '/path/to/r1.crt',
-            ],
-        ],
-    ],
-]);
