--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/mailer_with_smime_without_source.php
@@ -1,10 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'mailer' => [
-        'dsn' => 'smtp://example.com',
-        'smime_encrypter' => [
-            'enabled' => true,
-        ],
-    ],
-]);
