+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/legacy_mailer.yml
@@ -1,5 +1,3 @@
 framework:
     mailer:
         dsn: 'smtp://example.com'
-        smime_encrypter:
-            enabled: true
