--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_disabled_message_bus.yml
@@ -1,4 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        message_bus: false
