--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_dsn.yml
@@ -1,13 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        envelope:
-            sender: sender@example.org
-            recipients:
-                - redirected@example.org
-            allowed_recipients:
-                - foobar@example\.org
-        headers:
-            from: from@example.org
-            bcc: [bcc1@example.org, bcc2@example.org]
-            foo: bar
