--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_pgp_repository.yml
@@ -1,11 +1,0 @@
-services:
-    my_pgp_repository:
-        class: Symfony\Component\Mailer\EventListener\InMemoryPgpPublicKeyRepository
-        arguments: [[]]
-
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        pgp_encrypter:
-            enabled: true
-            repository: 'my_pgp_repository'
