--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_pgp_signer_without_secret_key.yml
@@ -1,5 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        pgp_signer:
-            enabled: true
