--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_pgp_without_source.yml
@@ -1,5 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        pgp_encrypter:
-            enabled: true
