--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_rate_limiter.yml
@@ -1,14 +1,0 @@
-framework:
-    rate_limiter:
-        foo_limiter:
-            lock_factory: ~
-            policy: token_bucket
-            limit: 10
-            rate:
-                interval: 5 seconds
-                amount: 10
-    mailer:
-        transports:
-            main:
-                dsn: smtp://example.com
-                rate_limiter: foo_limiter
