--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_smime_repository.yml
@@ -1,11 +1,0 @@
-services:
-    my_repository:
-        class: Symfony\Component\Mailer\EventListener\InMemorySmimeCertificateRepository
-        arguments: [[]]
-
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        smime_encrypter:
-            enabled: true
-            repository: 'my_repository'
