--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_smime_repository_and_certificates.yml
@@ -1,8 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        smime_encrypter:
-            enabled: true
-            repository: 'my_repository'
-            certificates:
-                r1@example.com: '/path/to/r1.crt'
