--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_tracking.yml
@@ -1,6 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        tracking:
-            opens: false
-            clicks: false
