--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_tracking_and_header.yml
@@ -1,8 +1,0 @@
-framework:
-    mailer:
-        dsn: 'smtp://example.com'
-        headers:
-            X-Track: 'opens=true; clicks=default'
-        tracking:
-            opens: false
-            clicks: false
