--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/mailer_with_transports.yml
@@ -1,17 +1,0 @@
-framework:
-    mailer:
-        transports:
-            transport1: 'smtp://example1.com'
-            transport2: 'smtp://example2.com'
-        envelope:
-            sender: sender@example.org
-            recipients:
-                - redirected@example.org
-                - redirected1@example.org
-            allowed_recipients:
-                - foobar@example\.org
-                - .*@example\.com
-        headers:
-            from: from@example.org
-            bcc: [bcc1@example.org, bcc2@example.org]
-            foo: bar
