+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -56,7 +56,7 @@
         "symfony/http-client": "^8.2",
         "symfony/json-streamer": "^8.2",
         "symfony/lock": "^8.2",
-        "symfony/mailer": "^7.4|^8.0",
+        "symfony/mailer": "^8.2",
         "symfony/messenger": "^8.2",
         "symfony/mime": "^8.2",
         "symfony/notifier": "^7.4|^8.0",
@@ -95,7 +95,7 @@
         "symfony/http-client": "<8.2",
         "symfony/json-path": "<8.2",
         "symfony/json-streamer": "<8.2",
-        "symfony/mailer": "<7.4.17|>=8.0,<8.1.5",
+        "symfony/mailer": "<8.2",
         "symfony/lock": "<8.2",
         "symfony/security-csrf": "<7.4",
         "symfony/semaphore": "<8.2",
