+++ src/Symfony/Component/Mailer/Resources/config/mailer.php
