+++ src/Symfony/Component/Mailer/Resources/config/mailer_transports.php
